Add shared feed-flow architecture and follows feed mode
