+++ BilibiliLive/Module/Tabbar/Notification+TabBar.swift
@@ -9,4 +9,5 @@ import Foundation
 
 extension Notification.Name {
     static let tabBarPagesDidChange = Notification.Name("BLTabBarViewController.tabBarPagesDidChange")
+    static let followsLayoutModeDidChange = Notification.Name("FollowsViewController.layoutModeDidChange")
 }
