+++ BilibiliLive/Module/ViewController/FollowsViewController.swift
@@ -6,10 +6,89 @@
 //
 
 import Alamofire
+import SnapKit
 import SwiftyJSON
 import UIKit
 
-class FollowsViewController: StandardVideoCollectionViewController<DynamicFeedData> {
+final class FollowsViewController: UIViewController, BLTabBarContentVCProtocol {
+    private enum LayoutMode {
+        case feedFlow
+        case grid
+    }
+
+    private var currentMode: LayoutMode?
+    private var currentContentViewController: UIViewController?
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .clear
+        NotificationCenter.default.addObserver(self,
+                                               selector: #selector(handleLayoutModeDidChange),
+                                               name: .followsLayoutModeDidChange,
+                                               object: nil)
+        syncLayoutIfNeeded(force: true)
+    }
+
+    deinit {
+        NotificationCenter.default.removeObserver(self)
+    }
+
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        syncLayoutIfNeeded(force: false)
+    }
+
+    override var preferredFocusEnvironments: [UIFocusEnvironment] {
+        currentContentViewController?.preferredFocusEnvironments ?? [view]
+    }
+
+    func reloadData() {
+        if let content = currentContentViewController as? BLTabBarContentVCProtocol {
+            content.reloadData()
+        } else {
+            syncLayoutIfNeeded(force: true)
+        }
+    }
+
+    @objc private func handleLayoutModeDidChange() {
+        guard isViewLoaded else { return }
+        syncLayoutIfNeeded(force: true)
+    }
+
+    private func syncLayoutIfNeeded(force: Bool) {
+        let targetMode: LayoutMode = Settings.followsFeedFlowEnabled ? .feedFlow : .grid
+        guard force || currentMode != targetMode else { return }
+
+        let targetViewController: UIViewController
+        switch targetMode {
+        case .feedFlow:
+            targetViewController = FollowsFeedFlowViewController()
+        case .grid:
+            targetViewController = FollowsGridViewController()
+        }
+
+        transition(to: targetViewController)
+        currentMode = targetMode
+    }
+
+    private func transition(to targetViewController: UIViewController) {
+        let previousViewController = currentContentViewController
+        addChild(targetViewController)
+        view.addSubview(targetViewController.view)
+        targetViewController.view.snp.makeConstraints { make in
+            make.edges.equalToSuperview()
+        }
+        targetViewController.didMove(toParent: self)
+        currentContentViewController = targetViewController
+
+        guard let previousViewController else { return }
+        previousViewController.willMove(toParent: nil)
+        previousViewController.view.removeFromSuperview()
+        previousViewController.removeFromParent()
+    }
+}
+
+final class FollowsGridViewController: StandardVideoCollectionViewController<DynamicFeedData> {
     var lastOffset = ""
 
     override func setupCollectionView() {
@@ -34,6 +113,79 @@ class FollowsViewController: StandardVideoCollectionViewController<DynamicFeedDa
     }
 }
 
+final class FollowsFeedFlowViewController: FeedFlowBrowserViewController {
+    init() {
+        super.init(dataSource: FollowsFeedFlowDataSource())
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
+final class FollowsFeedFlowDataSource: FeedFlowDataSource {
+    let title = "关注"
+    let defaultPreviewHintText = "停留后自动预览，按确认键进入视频流"
+    let loadingHintText = "正在加载关注视频..."
+    let emptyStateText = "当前关注区暂无可播放视频"
+    let emptyHintText = "可以在设置中关闭关注刷视频模式"
+    let loadFailureText = "关注加载失败，请稍后重试"
+    let refreshFailureHintText = "已展示现有关注内容，后台刷新失败"
+
+    var reloadToken: String {
+        "\(ApiRequest.getToken()?.mid ?? 0)"
+    }
+
+    private var lastOffset = ""
+    private var nextPage = 1
+    private var hasMore = true
+    private var items = [FeedFlowItem]()
+    private var seenItemKeys = Set<String>()
+
+    func reset() {
+        lastOffset = ""
+        nextPage = 1
+        hasMore = true
+        items = []
+        seenItemKeys = []
+    }
+
+    func refreshFromStart(targetCount: Int, maxSourcePages: Int) async throws -> [FeedFlowItem] {
+        let loadedItems = try await loadMoreUntilTarget(targetCount: targetCount, maxSourcePages: maxSourcePages)
+        items = loadedItems
+        return loadedItems
+    }
+
+    func loadMoreItems(targetCount: Int, maxSourcePages: Int) async throws -> [FeedFlowItem] {
+        let appended = try await loadMoreUntilTarget(targetCount: targetCount, maxSourcePages: maxSourcePages)
+        items.append(contentsOf: appended)
+        return appended
+    }
+
+    private func loadMoreUntilTarget(targetCount: Int, maxSourcePages: Int) async throws -> [FeedFlowItem] {
+        guard hasMore else { return [] }
+
+        var pagesScanned = 0
+        var accepted = [FeedFlowItem]()
+
+        while accepted.count < targetCount, pagesScanned < maxSourcePages, hasMore {
+            let info = try await WebRequest.requestFollowsFeed(offset: lastOffset, page: nextPage)
+            pagesScanned += 1
+            nextPage += 1
+            lastOffset = info.offset
+            hasMore = info.has_more
+
+            let newItems = info.videoFeeds
+                .compactMap(\.feedFlowItem)
+                .filter { seenItemKeys.insert($0.identityKey).inserted }
+            accepted.append(contentsOf: newItems)
+        }
+
+        return accepted
+    }
+}
+
 extension WebRequest {
     struct DynamicFeedInfo: Codable {
         let items: [DynamicFeedData]
@@ -41,9 +193,9 @@ extension WebRequest {
         let update_num: Int
         let update_baseline: String
         let has_more: Bool
+
         var videoFeeds: [DynamicFeedData] {
-            return items
-                .filter({ $0.aid != 0 || $0.modules.module_dynamic.major?.pgc != nil })
+            items.filter { $0.aid != 0 || $0.modules.module_dynamic.major?.pgc?.epid != nil }
         }
 
         enum CodingKeys: String, CodingKey {
@@ -71,7 +223,8 @@ extension WebRequest {
         if let offsetNum = Int(offset) {
             param["offset"] = offsetNum
         }
-        let res: DynamicFeedInfo = try await request(url: "https://api.bilibili.com/x/polymer/web-dynamic/v1/feed/all", parameters: param)
+        let res: DynamicFeedInfo = try await request(url: "https://api.bilibili.com/x/polymer/web-dynamic/v1/feed/all",
+                                                     parameters: param)
         if res.videoFeeds.isEmpty, res.has_more {
             return try await requestFollowsFeed(offset: res.offset, page: page)
         }
@@ -87,26 +240,26 @@ struct DynamicFeedData: Codable, PlayableData, DisplayData {
         return 0
     }
 
-    var cid: Int { return 0 }
+    var cid: Int { 0 }
 
     var title: String {
-        return modules.module_dynamic.major?.archive?.title ?? modules.module_dynamic.major?.pgc?.title ?? ""
+        modules.module_dynamic.major?.archive?.title ?? modules.module_dynamic.major?.pgc?.title ?? ""
     }
 
     var ownerName: String {
-        return modules.module_author.name
+        modules.module_author.name
     }
 
     var pic: URL? {
-        return URL(string: modules.module_dynamic.major?.archive?.cover ?? "") ?? modules.module_dynamic.major?.pgc?.cover
+        URL(string: modules.module_dynamic.major?.archive?.cover ?? "") ?? modules.module_dynamic.major?.pgc?.cover
     }
 
     var avatar: URL? {
-        return URL(string: modules.module_author.face)
+        URL(string: modules.module_author.face)
     }
 
     var date: String? {
-        return modules.module_author.pub_time
+        modules.module_author.pub_time
     }
 
     var overlay: DisplayOverlay? {
@@ -124,6 +277,34 @@ struct DynamicFeedData: Codable, PlayableData, DisplayData {
             rightItems.append(DisplayOverlay.DisplayOverlayItem(icon: nil, text: durationText))
         }
         return DisplayOverlay(leftItems: leftItems, rightItems: rightItems)
+    }
+
+    var feedFlowItem: FeedFlowItem? {
+        if aid > 0 {
+            return FeedFlowItem(aid: aid,
+                                title: title,
+                                ownerName: ownerName,
+                                coverURL: pic,
+                                avatarURL: avatar,
+                                duration: modules.module_dynamic.major?.archive?.duration_text?.durationInSeconds,
+                                durationText: modules.module_dynamic.major?.archive?.duration_text ?? "",
+                                viewCountText: modules.module_dynamic.major?.archive?.stat?.play ?? "",
+                                danmakuCountText: modules.module_dynamic.major?.archive?.stat?.danmaku ?? "",
+                                reasonText: date)
+        }
+
+        if let epid = modules.module_dynamic.major?.pgc?.epid, epid > 0 {
+            return FeedFlowItem(aid: 0,
+                                epid: epid,
+                                title: title,
+                                ownerName: ownerName,
+                                coverURL: pic,
+                                avatarURL: avatar,
+                                durationText: "",
+                                reasonText: date)
+        }
+
+        return nil
     }
 
     let type: String
@@ -194,5 +375,24 @@ struct DynamicFeedData: Codable, PlayableData, DisplayData {
                 }
             }
         }
+    }
+}
+
+private extension String {
+    var durationInSeconds: Int? {
+        let parts = split(separator: ":").compactMap { Int($0) }
+        guard !parts.isEmpty else { return nil }
+        var total = 0
+        for (index, value) in parts.reversed().enumerated() {
+            switch index {
+            case 0:
+                total += value
+            case 1:
+                total += value * 60
+            default:
+                total += value * 3600
+            }
+        }
+        return total
     }
 }
